+++ config/__init__.py
@@ -1,4 +1,0 @@
-"""Config package initialization."""
-from config.settings import config
-
-__all__ = ["config"]