+++ src/utils/__init__.py
@@ -1,4 +1,0 @@
-"""Utilities package initialization."""
-from src.utils.logger import get_logger
-
-__all__ = ["get_logger"]